Break up manager module a little bit
